move v2 as current version
