+++ cursor_handler.py
