--- v1/event_handler.py
@@ -1,7 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class EventHandler(ABC):
-    @abstractmethod
-    def react(self):
-        ...
