--- v1/test-project/session.json
@@ -1,19 +1,0 @@
-{
-  "image_path": "screen_shot.jpg",
-  "labels": [
-    {
-      "name": "fdas",
-      "type": "",
-      "label": "",
-      "text": "",
-      "top_left": [
-        -0.8145833333333334,
-        0.3833333333333334
-      ],
-      "bottom_right": [
-        -0.6567708333333334,
-        0.05277777777777792
-      ]
-    }
-  ]
-}
