--- v2/main.py
@@ -1,4 +1,0 @@
-from workspace import show_workspace, ProjectMenuEH
-
-# ProjectMenuEH().react()
-show_workspace('test-project')
