--- v2/requirements.txt
@@ -1,3 +1,0 @@
-PysimpleGUI~=4.60.1
-Pillow~=9.1.1
-PySimpleGUIQt~=0.35.0
