--- v2/test-project/export.json
@@ -1,212 +1,0 @@
-[
-  {
-    "name": "T-101",
-    "type": "Big Tank",
-    "label": "Raw Water Tank",
-    "text": "",
-    "coordinate": [
-      -0.75,
-      0.23611111111111116
-    ]
-  },
-  {
-    "name": "P-101",
-    "type": "Pump",
-    "label": "",
-    "text": "",
-    "coordinate": [
-      -0.565625,
-      0.2958333333333334
-    ]
-  },
-  {
-    "name": "P-102",
-    "type": "Pump",
-    "label": "Raw Water Pumps",
-    "text": "",
-    "coordinate": [
-      -0.5640624999999999,
-      0.125
-    ]
-  },
-  {
-    "name": "T-201",
-    "type": "Small Tank",
-    "label": "NaCl Tank",
-    "text": "",
-    "coordinate": [
-      -0.41093749999999996,
-      0.12777777777777788
-    ]
-  },
-  {
-    "name": "T-202",
-    "type": "Small Tank",
-    "label": "HCL Tank",
-    "text": "",
-    "coordinate": [
-      -0.315625,
-      0.13194444444444453
-    ]
-  },
-  {
-    "name": "T-203",
-    "type": "Small Tank",
-    "label": "NaCl Tank",
-    "text": "",
-    "coordinate": [
-      -0.21875,
-      0.12361111111111112
-    ]
-  },
-  {
-    "name": "T-301",
-    "type": "Big Tank",
-    "label": "UF Feed Water Tank",
-    "text": "",
-    "coordinate": [
-      0.06718749999999996,
-      0.0986111111111111
-    ]
-  },
-  {
-    "name": "P-301",
-    "type": "Pump",
-    "label": "",
-    "text": "",
-    "coordinate": [
-      0.23515625000000007,
-      0.17500000000000004
-    ]
-  },
-  {
-    "name": "P-302",
-    "type": "Pump",
-    "label": "UF Feed Pumps",
-    "text": "",
-    "coordinate": [
-      0.25234375,
-      0.015277777777777724
-    ]
-  },
-  {
-    "name": "T-401",
-    "type": "Big Tank",
-    "label": "RD Feed Water Tank",
-    "text": "",
-    "coordinate": [
-      0.634375,
-      0.24444444444444446
-    ]
-  },
-  {
-    "name": "P-401",
-    "type": "Pump",
-    "label": "",
-    "text": "",
-    "coordinate": [
-      0.8125,
-      0.30694444444444446
-    ]
-  },
-  {
-    "name": "P-402",
-    "type": "Pump",
-    "label": "RO Feed Pumps",
-    "text": "",
-    "coordinate": [
-      0.81328125,
-      0.14583333333333326
-    ]
-  },
-  {
-    "name": "UV-401",
-    "type": "UV Dechlorinator",
-    "label": "",
-    "text": "",
-    "coordinate": [
-      0.89453125,
-      -0.2888888888888889
-    ]
-  },
-  {
-    "name": "T-402",
-    "type": "Small Tank",
-    "label": "NaHSO3 Tank",
-    "text": "",
-    "coordinate": [
-      0.771875,
-      -0.46527777777777773
-    ]
-  },
-  {
-    "name": "Filter-501",
-    "type": "Filter",
-    "label": "",
-    "text": "",
-    "coordinate": [
-      0.6023437500000001,
-      -0.2972222222222222
-    ]
-  },
-  {
-    "name": "P-501",
-    "type": "Pump",
-    "label": "",
-    "text": "",
-    "coordinate": [
-      0.39531249999999996,
-      -0.23194444444444445
-    ]
-  },
-  {
-    "name": "P-502",
-    "type": "Pump",
-    "label": "RO Pumps",
-    "text": "",
-    "coordinate": [
-      0.396875,
-      -0.42083333333333334
-    ]
-  },
-  {
-    "name": "T-601",
-    "type": "Small Tank",
-    "label": "RO Permeate Tank",
-    "text": "",
-    "coordinate": [
-      -0.41328125,
-      -0.4875
-    ]
-  },
-  {
-    "name": "P-601",
-    "type": "Pump",
-    "label": "RO Permeate Pump",
-    "text": "",
-    "coordinate": [
-      -0.5296875,
-      -0.5055555555555555
-    ]
-  },
-  {
-    "name": "T-602",
-    "type": "Small Tank",
-    "label": "LF Backwash Tank",
-    "text": "",
-    "coordinate": [
-      -0.67578125,
-      -0.7402777777777778
-    ]
-  },
-  {
-    "name": "P-602",
-    "type": "Pump",
-    "label": "UF Backwash Pump",
-    "text": "",
-    "coordinate": [
-      -0.79921875,
-      -0.7847222222222222
-    ]
-  }
-]
