--- v2/test-project/test.json
@@ -1,32 +1,0 @@
-[
-  {
-    "name": "T-601",
-    "type": "",
-    "label": "",
-    "text": "",
-    "coordinate": [
-      -0.4125,
-      -0.4944444444444444
-    ]
-  },
-  {
-    "name": "",
-    "type": "",
-    "label": "",
-    "text": "",
-    "coordinate": [
-      -0.56640625,
-      0.12916666666666654
-    ]
-  },
-  {
-    "name": "",
-    "type": "",
-    "label": "",
-    "text": "",
-    "coordinate": [
-      0.07187500000000036,
-      0.12500000000000022
-    ]
-  }
-]
