--- v3/session.json
@@ -1,1 +1,0 @@
-{"image_path":
