+++ window_manager.py
